extract subcomponents as independent components under alert
