+++ src/components/Alert/stories/Alert.intro.stories.mdx
@@ -2,7 +2,15 @@ import { ArgsTable, Canvas, Meta } from '@storybook/addon-docs';
 
 import { Alert } from '../..';
 
-<Meta title="Components/Alert/Intro" component={Alert} />
+<Meta
+  title="Components/Alert/Intro"
+  component={Alert}
+  subcomponents={{
+    'Alert.Header': Alert.Header,
+    'Alert.Body': Alert.Body,
+    'Alert.Link': Alert.Link,
+  }}
+/>
 
 # Alert
